+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] / 2022-10-09
+### Features
+- Enable / Disable `ReleaseNameVersion` (default false)
+### Updated
+- Update `IHazRelease` add `ReleaseNameVersion`
+- Update `IHazRelease` add `GetReleaseFileNameVersion`
+- Update `IHazRelease` add `CreateReleaseFromDirectory`
+- Update `Build` to target "Build"
+- Update `BuildProject` and `RebuildProject`
+
 ## [1.3.6] / 2022-08-18
 ### Updated
 - Update GetGitRepositoryOwner null Exception
@@ -201,6 +211,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.4.0]: ../../compare/1.3.5...1.4.0
 [1.3.5]: ../../compare/1.3.4...1.3.5
 [1.3.4]: ../../compare/1.3.3...1.3.4
 [1.3.3]: ../../compare/1.3.2...1.3.3